Refactor Order and OrderItem DAOs
- Removed the OrderItemKey class from its previous location and redefined it in the order package.
- Implemented OrderDAO with methods for creating, reading, updating, and deleting orders.
- Added custom methods in OrderDAO for retrieving order summaries and detailed order information.
- Created OrderItemDAO to handle operations related to order items, including batch creation and retrieval.
- Updated database interaction methods to use prepared statements for better security and performance.
- Ensured proper handling of SQL exceptions and connection management throughout the DAOs.
