+++ backend/src/main/java/com/utephonehub/dao/order/OrderItemKey.java
@@ -1,4 +1,4 @@
-package com.utephonehub.dao;
+package com.utephonehub.dao.order;
 
 /**
  * Composite key class for OrderItem
